server/semantic_search/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pydantic
